Create select field creator class.
