+++ 2019-07-23-intro/app/view/index.js
@@ -1,4 +1,5 @@
 import {AnimalFactory} from "./animalfactory.js";
+import {SelectFieldCreator} from "./selectFieldCreator.js";
 
 const root = document.getElementById("root");
 let input = document.createElement("input");
@@ -8,55 +9,8 @@ input.style.marginRight = "50px";
 root.appendChild(input);
 
 let select = document.createElement("select");
+new SelectFieldCreator(select);
 
-
-let optionCat = document.createElement("option");
-optionCat.textContent = "Cat";
-
-let optionDog = document.createElement("option");
-optionDog.textContent = "Dog";
-
-let optionEagle = document.createElement("option");
-optionEagle.textContent = "Eagle";
-
-let optionElephant = document.createElement("option");
-optionElephant.textContent = "Elephant";
-
-let optionHorse = document.createElement("option");
-optionHorse.textContent = "Horse";
-
-let optionHuman = document.createElement("option");
-optionHuman.textContent = "Human";
-
-let optionLion = document.createElement("option");
-optionLion.textContent = "Lion";
-
-let optionParrot = document.createElement("option");
-optionParrot.textContent = "Parrot";
-
-let optionPig = document.createElement("option");
-optionPig.textContent = "Pig";
-
-let optionSalmon = document.createElement("option");
-optionSalmon.textContent = "Salmon";
-
-let optionTiger = document.createElement("option");
-optionTiger.textContent = "Tiger";
-
-
-select.appendChild(optionCat);
-
-select.appendChild(optionDog);
-select.appendChild(optionEagle);
-select.appendChild(optionElephant);
-select.appendChild(optionHorse);
-select.appendChild(optionHuman);
-select.appendChild(optionLion);
-select.appendChild(optionParrot);
-select.appendChild(optionPig);
-select.appendChild(optionSalmon);
-select.appendChild(optionTiger);
-select.selected = optionCat;
 
 root.appendChild(select);
 
@@ -73,12 +27,12 @@ displayButton.style.marginLeft = "50px";
 root.appendChild(displayButton);
 
 let animalFactory = new AnimalFactory();
-let animal;
+let animal = undefined;
 
 let list = document.createElement("ol");
 root.appendChild(list);
 
-addButton.addEventListener("click", event => {
+addButton.addEventListener("click", () => {
 
     let selectedAnimal = select.options[select.selectedIndex].textContent;
     animal = animalFactory.getAnimal(selectedAnimal);
@@ -96,7 +50,7 @@ addButton.addEventListener("click", event => {
 });
 
 
-displayButton.addEventListener("click", event => {
+displayButton.addEventListener("click", () => {
         let table = document.createElement("table");
         table.style.border = "1px solid black";
         root.appendChild(table);
